Invincibility logic added
Healthpacks now have a rigidbody component for collision. Healthpack
collision boundary is too large however.

Invincibility is currently global for all players. If one player touches
a healthpack ALL players become invincible for 10 seconds. Timer needs
to be separated between the four players in the future.

KillPlayer now contains a boolean variable so that it can check whether
the player was killed by the environment or the boss.