Layered BGM
Here are some tracks for the level BGM. The idea is that the first track
is the base music for the level, then track 2 is some kickin' synth
soloing, track 3 is some super cheesy brass and hyperactive extra drums,
and track 4 adds in some epic choir and strings for the final part of
the level.

Was thinking it might be good to use the original BGM for when the boss
was about to win?